✨Knowledge_base summary:change summary func name
