+++ frontend/services/api.ts
@@ -55,8 +55,8 @@ export const API_ENDPOINTS = {
     indices: `${API_BASE_URL}/indices`,
     indexInfo: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/info`,
     indexDetail: (indexName: string) => `${API_BASE_URL}/indices/${indexName}`,
-    summary: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/summary`,
-    changeSummary: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/change_summary`,
+    autoSummary: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/auto_summary`,
+    changeSummary: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/summary`,
     getSummary: (indexName: string) => `${API_BASE_URL}/indices/${indexName}/summary`,
     
     // 文件上传服务
